+++ src/app/admin/institution-settings/overseer-images/overseer-image-list.component.ts
@@ -89,7 +89,9 @@ export class OverseerImageListComponent
   // Push the values that will be displayed in the table
   // to the datasource
   private pushToTable(value: OverseerImage | OverseerImage[]) {
-    if (!value) return;
+    if (!value) {
+      return;
+    }
     if (value instanceof Array) {
       this.overseerImages.push(...value);
     } else {