+++ src/app/common/filters/tasks-by-tutor.pipe.ts
@@ -7,12 +7,18 @@ import {Task, UnitRole} from '../../api/models/doubtfire-model';
 })
 export class TasksByTutorPipe implements PipeTransform {
   transform(currentUnitRole: UnitRole, tasks: Task[], unitRoleId?: number | string): Task[] {
-    if (!tasks) return tasks;
+    if (!tasks) {
+      return tasks;
+    }
 
-    if (!unitRoleId || unitRoleId === 'all') return tasks;
+    if (!unitRoleId || unitRoleId === 'all') {
+      return tasks;
+    }
 
     if (unitRoleId === 'mentoring_all') {
-      if (!currentUnitRole) return [];
+      if (!currentUnitRole) {
+        return [];
+      }
 
       return tasks.filter((task) => task.tutor?.mentorId === currentUnitRole.id);
     }