+++ src/app/common/filters/tasks-for-group-set.pipe.ts
@@ -7,7 +7,9 @@ import {GroupSet, Task} from 'src/app/api/models/doubtfire-model';
 })
 export class TasksForGroupsetPipe implements PipeTransform {
   transform(tasks: readonly Task[], groupSet: GroupSet): Task[] {
-    if (!tasks) return [];
+    if (!tasks) {
+      return [];
+    }
 
     return tasks.filter((task) => {
       return task.definition.groupSet === groupSet || (!task.definition.groupSet && !groupSet);